Extract HttpClient creation to the platform
