+++ shared/src/androidMain/kotlin/org/hyperskill/app/network/PreconfiguredHttpClient.kt
@@ -1,0 +1,11 @@
+package org.hyperskill.app.network
+
+import io.ktor.client.HttpClient
+import io.ktor.client.HttpClientConfig
+import io.ktor.client.engine.okhttp.OkHttp
+
+@Suppress("FunctionName")
+internal actual fun PreconfiguredHttpClient(block: HttpClientConfig<*>.() -> Unit): HttpClient =
+    HttpClient(OkHttp) {
+        apply(block)
+    }
