+++ shared/src/iosMain/kotlin/org/hyperskill/app/network/PreconfiguredHttpClient.kt
@@ -1,0 +1,11 @@
+package org.hyperskill.app.network
+
+import io.ktor.client.HttpClient
+import io.ktor.client.HttpClientConfig
+import io.ktor.client.engine.darwin.Darwin
+
+@Suppress("FunctionName")
+internal actual fun PreconfiguredHttpClient(block: HttpClientConfig<*>.() -> Unit): HttpClient =
+    HttpClient(Darwin) {
+        apply(block)
+    }
